+++ src/database/Mongo.kt
@@ -1,4 +1,4 @@
-package net.tkhamez.everoute
+package net.tkhamez.everoute.database
 
 import net.tkhamez.everoute.data.MongoAlliance
 import net.tkhamez.everoute.data.MongoAnsiblex
@@ -6,12 +6,12 @@ import net.tkhamez.everoute.data.MongoTemporaryConnection
 import org.litote.kmongo.*
 import java.util.*
 
-class Mongo(uri: String) {
+class Mongo(uri: String): DbInterface {
     private val dbName = uri.substring(uri.lastIndexOf('/') + 1)
     private val client = KMongo.createClient(uri)
     private val database = client.getDatabase(dbName)
 
-    fun gatesGet(allianceId: Int): List<MongoAnsiblex> {
+    override fun gatesGet(allianceId: Int): List<MongoAnsiblex> {
         val gates = mutableListOf<MongoAnsiblex>()
         database.getCollection<MongoAnsiblex>("ansiblex")
             .find(MongoAnsiblex::allianceId eq allianceId)
@@ -19,7 +19,7 @@ class Mongo(uri: String) {
         return gates
     }
 
-    fun gateStore(ansiblex: MongoAnsiblex, allianceId: Int) {
+    override fun gateStore(ansiblex: MongoAnsiblex, allianceId: Int) {
         val col = database.getCollection<MongoAnsiblex>("ansiblex")
         val filter = and(
             MongoAnsiblex::id eq ansiblex.id,
@@ -33,7 +33,7 @@ class Mongo(uri: String) {
         }
     }
 
-    fun gatesRemoveOther(ansiblexes: List<MongoAnsiblex>, allianceId: Int) {
+    override fun gatesRemoveOther(ansiblexes: List<MongoAnsiblex>, allianceId: Int) {
         val col = database.getCollection<MongoAnsiblex>("ansiblex")
 
         val ids: MutableList<Long> = mutableListOf()
@@ -45,7 +45,7 @@ class Mongo(uri: String) {
         }
     }
 
-    fun temporaryConnectionsGet(characterId: Int): List<MongoTemporaryConnection> {
+    override fun temporaryConnectionsGet(characterId: Int): List<MongoTemporaryConnection> {
         val col = database.getCollection<MongoTemporaryConnection>("temporary-connection")
 
         val existingConnections = mutableListOf<MongoTemporaryConnection>()
@@ -56,7 +56,7 @@ class Mongo(uri: String) {
         return existingConnections
     }
 
-    fun temporaryConnectionStore(tempConnection: MongoTemporaryConnection) {
+    override fun temporaryConnectionStore(tempConnection: MongoTemporaryConnection) {
         val col = database.getCollection<MongoTemporaryConnection>("temporary-connection")
         val filter = and(
             MongoTemporaryConnection::system1Id eq tempConnection.system1Id,
@@ -71,14 +71,14 @@ class Mongo(uri: String) {
         }
     }
 
-    fun temporaryConnectionsDeleteAllExpired() {
+    override fun temporaryConnectionsDeleteAllExpired() {
         val col = database.getCollection<MongoTemporaryConnection>("temporary-connection")
         val calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"))
         calendar.add(Calendar.DAY_OF_YEAR, -2)
         col.deleteMany(MongoTemporaryConnection::created lt calendar.time)
     }
 
-    fun temporaryConnectionDelete(system1Id: Int, system2Id: Int, characterId: Int) {
+    override fun temporaryConnectionDelete(system1Id: Int, system2Id: Int, characterId: Int) {
         val col = database.getCollection<MongoTemporaryConnection>("temporary-connection")
         col.deleteOne(
             MongoTemporaryConnection::system1Id eq system1Id,
@@ -87,12 +87,12 @@ class Mongo(uri: String) {
         )
     }
 
-    fun allianceGet(allianceId: Int): MongoAlliance? {
+    override fun allianceGet(allianceId: Int): MongoAlliance? {
         val col = database.getCollection<MongoAlliance>("alliance")
         return col.findOne(MongoAlliance::id eq allianceId)
     }
 
-    fun allianceUpdate(alliance: MongoAlliance) {
+    override fun allianceUpdate(alliance: MongoAlliance) {
         val col = database.getCollection<MongoAlliance>("alliance")
         val existingAlliance = col.findOne(MongoAlliance::id eq alliance.id)
         if (existingAlliance == null) {