+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 yyyy-mm-dd
 
 - Fixed some MongoDB operations.
+- [#3][i3] Added support for PostgreSQL, MySQL, MariaDB, SQLite and H2 (embedded mode) databases.
+
+[i3]: https://github.com/tkhamez/eve-route/issues/3
 
 ## 0.1.0